Refactor firebase storage code
