+++ src/camera/actions.js
@@ -1,5 +1,4 @@
-import firebase from "firebase";
-import moment from "moment";
+import { uploadBase64Image } from "../firebase/firebaseStorage";
 
 export const SAVE_TENTATIVE_PICTURE = "SAVE_TENTATIVE_PICTURE";
 export const UPLOAD_PICTURE = "UPLOAD_PICTURE";
@@ -19,21 +18,8 @@ export function uploadPicture(base64Picture) {
       base64Picture: base64Picture
     });
     const userEmail = getState().session.user.email;
-    const storageRef = firebase.storage().ref();
-    const timestamp = moment().valueOf();
-    const fileExtension = "jpg";
-    const childRef = storageRef.child(
-      `${userEmail}/${timestamp}.${fileExtension}`
-    );
-    childRef.putString(base64Picture, "data_url").then(() => {
-      childRef.getDownloadURL().then(resourceUrl => {
-        const pictureUrl =
-          resourceUrl.substring(
-            0,
-            resourceUrl.indexOf(fileExtension) + fileExtension.length
-          ) + "?alt=media";
-        dispatch(pictureUploaded(pictureUrl));
-      });
+    uploadBase64Image(userEmail, base64Picture).then(pictureUrl => {
+      dispatch(pictureUploaded(pictureUrl));
     });
   };
 }
